+++ code/addons/interactions/src/components/EmptyState.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import { Link, EmptyTabContent } from '@storybook/components';
+import { Link, EmptyTabContent } from 'storybook/internal/components';
 import { DocumentIcon, VideoIcon } from '@storybook/icons';
-import { useStorybookApi } from '@storybook/manager-api';
-import { styled } from '@storybook/theming';
+import { useStorybookApi } from 'storybook/internal/manager-api';
+import { styled } from 'storybook/internal/theming';
 
 import { DOCUMENTATION_LINK, TUTORIAL_VIDEO_LINK } from '../constants';