+++ code/lib/cli/src/generators/ANGULAR/helpers.ts
@@ -1,10 +1,10 @@
 import fs from 'fs';
 import { join } from 'path';
 import prompts from 'prompts';
-import dedent from 'ts-dedent';
-import { MissingAngularJsonError } from '@storybook/core-events/server-errors';
+import { dedent } from 'ts-dedent';
+import { MissingAngularJsonError } from 'storybook/internal/server-errors';
 import boxen from 'boxen';
-import { logger } from '@storybook/node-logger';
+import { logger } from '@storybook/core/node-logger';
 
 export const ANGULAR_JSON_PATH = 'angular.json';