+++ code/lib/codemod/src/transforms/__tests__/mdx-to-csf.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, expect, vi, it } from 'vitest';
 import * as fs_ from 'node:fs';
-import dedent from 'ts-dedent';
+import { dedent } from 'ts-dedent';
 import jscodeshift, { nameToValidExport } from '../mdx-to-csf';
 
 expect.addSnapshotSerializer({