+++ code/lib/codemod/src/transforms/__tests__/migrate-to-test-package.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest';
 import transform from '../migrate-to-test-package';
-import dedent from 'ts-dedent';
+import { dedent } from 'ts-dedent';
 
 expect.addSnapshotSerializer({
   serialize: (val: any) => (typeof val === 'string' ? val : val.toString()),